+++ src/android/com/outsystems/plugins/healthfitness/OSHealthFitness.kt
@@ -178,36 +178,52 @@ class OSHealthFitness : CordovaImplementation() {
     }
 
     private fun writeData(args: JSONArray) {
-
-        //process parameters
         val variable = args.getString(0)
-        val value = args.getDouble(1).toFloat()
+        val value = args.getDouble(1)
+        val healthRecordName: HealthRecordName
 
-        healthStore?.updateDataAsync(
-            variable,
-            value,
-            { response ->
-                sendPluginResult(response)
-            },
-            { error ->
-                sendPluginResult(null, Pair(error.code.toString(), error.message))
+        when (variable) {
+            HealthRecordName.WEIGHT.name -> {
+                healthRecordName = HealthRecordName.WEIGHT
             }
+            HealthRecordName.HEIGHT.name -> {
+                healthRecordName = HealthRecordName.HEIGHT
+            }
+            HealthRecordName.BODY_FAT_PERCENTAGE.name -> {
+                healthRecordName = HealthRecordName.BODY_FAT_PERCENTAGE
+            }
+            HealthRecordName.BASAL_METABOLIC_RATE.name -> {
+                healthRecordName = HealthRecordName.BASAL_METABOLIC_RATE
+            }
+            else -> {
+                sendPluginResult(
+                    null,
+                    Pair(
+                        HealthFitnessError.WRITE_DATA_NOT_PROFILE_ERROR.code.toString(),
+                        HealthFitnessError.WRITE_DATA_NOT_PROFILE_ERROR.message
+                    )
+                )
+                return
+            }
+        }
 
+        healthConnectViewModel.writeData(
+            healthRecordName,
+            value,
+            getActivity().packageName,
+            {
+                sendPluginResult("success", null)
+            },
+            {
+                sendPluginResult(null, Pair(it.code.toString(), it.message))
+            }
         )
     }
 
     private fun getLastRecord(args: JSONArray) {
         //process parameters
         val variable = args.getString(0)
-        healthStore?.getLastRecordAsync(
-            variable,
-            { response ->
-                val pluginResponseJson = gson.toJson(response)
-                sendPluginResult(pluginResponseJson)
-            },
-            { error ->
-                sendPluginResult(null, Pair(error.code.toString(), error.message))
-            })
+        healthConnectViewModel.getLastRecord()
     }
 
     private fun setBackgroundJob(args: JSONArray) {