+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The changes documented here do not include those from the original repository.
 
 ## [Unreleased]
 
-## 2023-08-25
+## 2024-02-01
 - Re-implemented RequestPermissions feature (https://outsystemsrd.atlassian.net/browse/RMET-3046).
 
 ## [Version 1.4.0]