+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The changes documented here do not include those from the original repository.
 
 ## [Unreleased]
 
+## 2024-02-05
+- Re-implemented WriteProfieleData feature (https://outsystemsrd.atlassian.net/browse/RMET-3049).
+
 ## 2024-02-01
 - Re-implemented RequestPermissions feature (https://outsystemsrd.atlassian.net/browse/RMET-3046).