+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The changes documented here do not include those from the original repository.
 
 ## [Unreleased]
 
+ ## 2024-04-05
+- Fix privacy policy opening for Android <= 13 (https://outsystemsrd.atlassian.net/browse/RMET-3134).
+
  ## 2024-04-04
 - Update GSON version to remove vulnerability (https://outsystemsrd.atlassian.net/browse/RMET-3134).