+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 The changes documented here do not include those from the original repository.
 
+## [Unreleased]
+
 ## [Version 1.2.11]
 - Fix: [iOS] Replace the old `OSCore` framework for the new `OSCommonPluginLib` pod.